+++ src/snowflake/cli/api/rest_api.py
@@ -155,8 +155,9 @@ class RestApi:
             raise SchemaNotDefinedException(
                 "Schema not defined in connection. Please try again with `--schema` flag."
             )
-        if not self._schema_exists(db_name=db, schema_name=schema):
-            raise SchemaNotExistsException(f"Schema '{schema}' does not exist.")
+        # temporarily disable this check due to an issue on server side: SNOW-1747450
+        # if not self._schema_exists(db_name=db, schema_name=schema):
+        #     raise SchemaNotExistsException(f"Schema '{schema}' does not exist.")
         if self.get_endpoint_exists(
             url := f"{SF_REST_API_URL_PREFIX}/databases/{self.conn.database}/schemas/{self.conn.schema}/{plural_object_type}/"
         ):