+++ tests/nativeapp/test_application_package_entity.py
@@ -51,8 +51,10 @@ from tests.nativeapp.utils import (
 )
 
 
-def _get_app_pkg_entity(project_directory, package_overrides=None):
-    with project_directory("workspaces_simple") as project_root:
+def _get_app_pkg_entity(
+    project_directory, test_dir="workspaces_simple", package_overrides=None
+):
+    with project_directory(test_dir) as project_root:
         with Path(project_root / "snowflake.yml").open() as definition_file_path:
             project_definition = yaml.safe_load(definition_file_path)
             project_definition["entities"]["pkg"] = dict(